Rename pages correctly
