+++ lib/presentation/in_exp_history_widget.dart
@@ -1,14 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
+import 'package:shmr_finance/app_theme.dart';
+import 'package:shmr_finance/domain/cubit/datepicker_cubit.dart';
+import 'package:shmr_finance/domain/cubit/transaction_cubit.dart';
 import 'package:shmr_finance/presentation/widgets/custom_appbar.dart';
 import 'package:shmr_finance/presentation/widgets/item_inexp.dart';
 
-import '../app_theme.dart';
-import '../domain/cubit/transaction_cubit.dart';
-
 class InExpHistoryWidget extends StatefulWidget {
   final bool isIncome;
+
   const InExpHistoryWidget({super.key, required this.isIncome});
 
   @override
@@ -17,12 +18,40 @@ class InExpHistoryWidget extends StatefulWidget {
 
 class _InExpHistoryWidgetState extends State<InExpHistoryWidget> {
   @override
+  void initState() {
+    super.initState();
+
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      final datePickerState = context.read<DatePickerCubit>().state;
+      context.read<TransactionCubit>().fetchTransactions(
+        startDate: datePickerState.startDate!,
+        endDate: datePickerState.endDate!,
+        isIncome: widget.isIncome,
+      );
+    });
+  }
+
+  @override
+  void didUpdateWidget(covariant InExpHistoryWidget oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (oldWidget.isIncome != widget.isIncome) {
+      final datePickerState = context.read<DatePickerCubit>().state;
+      context.read<TransactionCubit>().fetchTransactions(
+        startDate: datePickerState.startDate!,
+        endDate: datePickerState.endDate!,
+        isIncome: widget.isIncome,
+      );
+    }
+  }
+
+  @override
   Widget build(BuildContext context) {
+    DatePickerCubit datePickerCubit = context.read<DatePickerCubit>();
     TransactionCubit transactionCubit = context.read<TransactionCubit>();
 
     return Scaffold(
       appBar: CustomAppBar(
-        title: widget.isIncome ? "Доходы сегодня" : "Расходы сегодня",
+        title: widget.isIncome ? "История доходов" : "История расходов",
         actions: <Widget>[
           IconButton(
             icon: const Icon(Icons.analytics_outlined),
@@ -31,7 +60,7 @@ class _InExpHistoryWidgetState extends State<InExpHistoryWidget> {
                 context,
                 MaterialPageRoute<void>(
                   builder: (BuildContext context) {
-                    return Placeholder(); // TODO: передавать с параметром isIncome
+                    return Placeholder();
                   },
                 ),
               );
@@ -39,82 +68,200 @@ class _InExpHistoryWidgetState extends State<InExpHistoryWidget> {
           ),
         ],
       ),
-      body: BlocBuilder<TransactionCubit, TransactionState>(
-        builder: (context, state) {
-          transactionCubit.fetchTransactions(
-            startDate: DateTime.now().copyWith(hour: 0, minute: 0),
-            endDate: DateTime.now().copyWith(hour: 23, minute: 59),
-            isIncome: widget.isIncome,
-          );
-          final transactions = state.transactions;
-          final totalSum = transactions.fold<num>(
-            0,
+      body: BlocBuilder<DatePickerCubit, DatePickerState>(
+        builder: (context, datePickerState) {
+          return BlocBuilder<TransactionCubit, TransactionState>(
+            builder: (context, transactionState) {
+              final transactions = transactionState.transactions;
+              final totalSum = transactions.fold<num>(
+                0,
                 (sum, item) => sum + double.parse(item.amount),
-          );
-          return Column(
-            children: [
-              Container(
-                color: CustomAppTheme.figmaMainLightColor,
-                height: 56,
-                child: Row(
-                  children: [
-                    const Expanded(
-                      child: Padding(
-                        padding: EdgeInsets.only(left: 16),
-                        child: Text("Всего", textAlign: TextAlign.start),
+              );
+              return Column(
+                children: [
+                  // Начало
+                  GestureDetector(
+                    onTap: () async {
+                      final DateTime? resultStartDate = await showDatePicker(
+                        context: context,
+                        initialDate:
+                            datePickerState.startDate ?? DateTime.now(),
+                        firstDate: DateTime(2000),
+                        lastDate: DateTime(2100),
+                        helpText: 'Выберите дату начала',
+                        cancelText: 'Отмена',
+                        confirmText: 'ОК',
+                      );
+                      if (resultStartDate != null) {
+                        datePickerCubit.setStartDate(resultStartDate);
+                        transactionCubit.fetchTransactions(
+                          startDate: resultStartDate,
+                          endDate: datePickerState.endDate,
+                          isIncome: widget.isIncome,
+                        );
+                      }
+                    },
+                    child: Container(
+                      color: CustomAppTheme.figmaMainLightColor,
+                      height: 56,
+                      child: Row(
+                        children: [
+                          const Expanded(
+                            child: Padding(
+                              padding: EdgeInsets.only(left: 16),
+                              child: Text("Начало", textAlign: TextAlign.start),
+                            ),
+                          ),
+                          Expanded(
+                            child: Padding(
+                              padding: const EdgeInsets.only(right: 16),
+                              child: Text(
+                                datePickerState.startDate != null
+                                    ? DateFormat(
+                                      'dd.MM.yyyy',
+                                    ).format(datePickerState.startDate!)
+                                    : "Выберите дату",
+                                textAlign: TextAlign.end,
+                              ),
+                            ),
+                          ),
+                        ],
                       ),
                     ),
-                    Expanded(
-                      child: Padding(
-                        padding: const EdgeInsets.only(right: 16),
-                        child: Text("${NumberFormat("0.00").format(totalSum)} ₽", textAlign: TextAlign.end),
+                  ),
+                  const Divider(
+                    height: 1,
+                    thickness: 1,
+                    color: CustomAppTheme.figmaBgGrayColor,
+                  ),
+                  // Конец
+                  GestureDetector(
+                    onTap: () async {
+                      final DateTime? resultEndDate = await showDatePicker(
+                        context: context,
+                        initialDate: datePickerState.endDate ?? DateTime.now(),
+                        firstDate: DateTime(2000),
+                        lastDate: DateTime(2100),
+                        helpText: 'Выберите дату конца',
+                        cancelText: 'Отмена',
+                        confirmText: 'ОК',
+                      );
+                      if (resultEndDate != null) {
+                        datePickerCubit.setEndDate(resultEndDate);
+                      }
+                    },
+                    child: Container(
+                      color: CustomAppTheme.figmaMainLightColor,
+                      height: 56,
+                      child: Row(
+                        children: [
+                          const Expanded(
+                            child: Padding(
+                              padding: EdgeInsets.only(left: 16),
+                              child: Text("Конец", textAlign: TextAlign.start),
+                            ),
+                          ),
+                          Expanded(
+                            child: Padding(
+                              padding: const EdgeInsets.only(right: 16),
+                              child: Text(
+                                datePickerState.endDate != null
+                                    ? DateFormat(
+                                      'dd.MM.yyyy',
+                                    ).format(datePickerState.endDate!)
+                                    : "Выберите дату",
+                                textAlign: TextAlign.end,
+                              ),
+                            ),
+                          ),
+                        ],
                       ),
                     ),
-                  ],
-                ),
-              ),
-              Expanded(
-                child: () {
-                  if (state.status == TransactionStatus.loading) {
-                    return const Center(child: CircularProgressIndicator());
-                  }
-                  if (state.status == TransactionStatus.error) {
-                    return Center(child: Text('Ошибка: ${state.error}'));
-                  }
-                  if (transactions.isEmpty) {
-                    return const Center(child: Text('Нет данных за период'));
-                  } else {
-                    return ListView.builder(
-                      itemCount: transactions.length * 2 + 1,
-                      itemBuilder: (context, index) {
-                        if (index.isEven) {
-                          return const Divider(
-                            height: 1,
-                            thickness: 1,
-                            color: CustomAppTheme.figmaBgGrayColor,
-                          );
-                        } else {
-                          final itemIndex = index ~/ 2;
-                          if (itemIndex >= transactions.length) {
-                            return const SizedBox.shrink();
-                          }
-                          final item = transactions[itemIndex];
-                          return InExpItem(
-                            categoryTitle: item.category.name,
-                            amount: item.amount,
-                            icon: item.category.emoji,
-                            time: item.transactionDate,
-                            comment: item.comment,
-                          );
-                        }
-                      },
-                    );
-                  }
-                }(),
-              ),
-            ],
+                  ),
+                  const Divider(
+                    height: 1,
+                    thickness: 1,
+                    color: CustomAppTheme.figmaBgGrayColor,
+                  ),
+                  // Сумма
+                  Container(
+                    color: CustomAppTheme.figmaMainLightColor,
+                    height: 56,
+                    child: Row(
+                      children: [
+                        const Expanded(
+                          child: Padding(
+                            padding: EdgeInsets.only(left: 16),
+                            child: Text("Сумма", textAlign: TextAlign.start),
+                          ),
+                        ),
+                        Expanded(
+                          child: Padding(
+                            padding: const EdgeInsets.only(right: 16),
+                            child: Text(
+                              "$totalSum ₽", // TODO: добавить форматирование вывода суммы по разрядам и кол-во символов после точки
+                              textAlign: TextAlign.end,
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                  // Список транзакций
+                  Expanded(
+                    child: () {
+                      if (transactionState.status ==
+                          TransactionStatus.loading) {
+                        return const Center(child: CircularProgressIndicator());
+                      }
+                      if (transactionState.status == TransactionStatus.error) {
+                        return Center(
+                          child: Text('Ошибка: ${transactionState.error}'),
+                        );
+                      }
+                      if (transactions.isEmpty) {
+                        return const Center(
+                          child: Text('Нет данных за период'),
+                        );
+                      } else {
+                        return ListView.builder(
+                          itemCount: transactions.length * 2 + 1,
+                          itemBuilder: (context, index) {
+                            if (index.isEven) {
+                              return const Divider(
+                                height: 1,
+                                thickness: 1,
+                                color: CustomAppTheme.figmaBgGrayColor,
+                              );
+                            } else {
+                              final itemIndex = index ~/ 2;
+                              if (itemIndex >= transactions.length) {
+                                return const SizedBox.shrink();
+                              }
+                              final item = transactions[itemIndex];
+                              return InExpItem(
+                                categoryTitle: item.category.name,
+                                amount: item.amount,
+                                icon: item.category.emoji,
+                                time: item.transactionDate,
+                                comment: item.comment,
+                              );
+                            }
+                          },
+                        );
+                      }
+                    }(),
+                  ),
+                ],
+              );
+            },
           );
         },
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {},
+        shape: CircleBorder(),
+        child: Icon(Icons.add),
       ),
     );
   }
